pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "routerbot"
version = "1.0.0"
description = "Open Source LLM Gateway — Call 100+ LLMs through a single, unified API. No enterprise paywalls."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "RouterBot Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: FastAPI",
    "Typing :: Typed",
]
keywords = ["llm", "openai", "anthropic", "gateway", "proxy", "ai"]

dependencies = [
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.1,<3",
    "httpx>=0.27,<1",
    "tiktoken>=0.7,<1",
    "structlog>=24.1,<25",
    "pyyaml>=6.0,<7",
    "orjson>=3.9,<4",
    "tenacity>=8.2,<10",
    "click>=8.1,<9",
    "prometheus-client>=0.24.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
]

[project.optional-dependencies]
proxy = [
    "fastapi>=0.111,<1",
    "uvicorn[standard]>=0.29,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.29,<1",
    "aiosqlite>=0.20,<1",
    "alembic>=1.13,<2",
    "redis[hiredis]>=5.0,<6",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3,<4",
    "passlib[bcrypt]>=1.7,<2",
    "aiofiles>=23.2,<25",
]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "pytest-xdist>=3.5,<4",
    "pytest-timeout>=2.2,<3",
    "respx>=0.21,<1",
    "factory-boy>=3.3,<4",
    "ruff>=0.4,<1",
    "mypy>=1.9,<2",
    "types-PyYAML>=6.0",
    "pre-commit>=3.6,<4",
    "pip-audit>=2.7,<3",
]

[project.urls]
Homepage = "https://github.com/Nivesh30/routerbot"
Repository = "https://github.com/Nivesh30/routerbot"
Documentation = "https://github.com/Nivesh30/routerbot/tree/master/docs"
Issues = "https://github.com/Nivesh30/routerbot/issues"

[project.scripts]
routerbot = "routerbot.cli:main"

# ─── Tool Configuration ────────────────────────────────────

[tool.hatch.build.targets.wheel]
packages = ["src/routerbot"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T20",   # flake8-print
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
    "S",     # flake8-bandit (security)
    "ASYNC", # flake8-async
]
ignore = [
    "E501",   # line length handled by formatter
    "S101",   # allow assert in tests
    "B008",   # allow function calls in default args (FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["routerbot"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "tiktoken.*", "jose.*", "passlib.*", "yaml.*",
    "boto3.*", "botocore.*",
    "azure.*",
    "google.cloud.*", "google.auth.*",
    "redis.*", "aiofiles.*",
    "langfuse.*", "opentelemetry.*",
    "fastapi.*", "uvicorn.*", "starlette.*",
    "sqlalchemy.*", "alembic.*", "asyncpg.*", "aiosqlite.*",
    "respx.*", "httpx.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "routerbot.core.secrets.*",
    "routerbot.observability.exporters.*",
    "routerbot.observability.opentelemetry",
    "routerbot.core.plugins.examples.*",
    "routerbot.core.plugins.manager",
    "routerbot.cache.redis",
    "routerbot.cache.manager",
    "routerbot.proxy.app",
    "routerbot.proxy.middleware.ip_filter",
    "routerbot.proxy.routes.dashboard",
    "routerbot.proxy.routes.model_management",
    "routerbot.core.semantic.classifier",
    "routerbot.core.mcp.client",
    "routerbot.core.resilience.circuit_breaker",
    "routerbot.auth.advanced.token_exchange",
    "routerbot.k8s.crd_schemas",
    "routerbot.evaluation.regression",
]
disallow_untyped_defs = false
ignore_errors = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
ignore_errors = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-v",
    "--tb=short",
]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (requires Docker services)",
    "slow: Slow tests (excluded from default run)",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    # Windows ProactorEventLoop: orphaned transport/socket __del__ warnings
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

[tool.coverage.run]
source = ["routerbot"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__:",
    "@overload",
    "raise NotImplementedError",
    "\\.\\.\\.",
]
show_missing = true
fail_under = 80
